Move demo externalCommand to separate file
