+++ demo/serve/external_command.ts
@@ -1,0 +1,53 @@
+import { ansi, ExitCode, IExternalContext } from '@jupyterlite/cockle';
+
+export async function externalCommand(context: IExternalContext): Promise<number> {
+  const { args } = context;
+
+  if (args.includes('environment')) {
+    context.environment.set('TEST_VAR', '23');
+    context.environment.delete('TEST_VAR2');
+  }
+
+  if (args.includes('name')) {
+    context.stdout.write(context.name + '\n');
+  }
+
+  if (args.includes('stdout')) {
+    context.stdout.write('Output line 1\n');
+    context.stdout.write('Output line 2\n');
+  }
+
+  if (args.includes('stderr')) {
+    context.stderr.write('Error message\n');
+  }
+
+  if (args.includes('color')) {
+    for (let j = 0; j < 16; j++) {
+      let line = '';
+      for (let i = 0; i < 32; i++) {
+        const rgb = ansi.styleRGB((i + 1) * 8 - 1, 128, (j + 1) * 16 - 1);
+        line += rgb + String.fromCharCode(65 + i) + ansi.styleReset;
+      }
+      context.stdout.write(line + '\n');
+    }
+  }
+
+  if (args.includes('stdin')) {
+    // Read until EOT, echoing back as upper case.
+    const { stdin, stdout } = context;
+    let stop = false;
+    while (!stop) {
+      const chars = await stdin.readAsync(null);
+      if (chars.length === 0 || chars.endsWith('\x04')) {
+        stop = true;
+      } else {
+        stdout.write(chars.toUpperCase());
+      }
+    }
+  }
+
+  if (args.includes('exitCode')) {
+    return ExitCode.GENERAL_ERROR;
+  }
+  return ExitCode.SUCCESS;
+}
